+++ R/utils.R
@@ -1,3 +1,10 @@
+ensure_daemons <- function(envir = parent.frame()) {
+  if (status()$connections == 0L) {
+    daemons(parallel::detectCores())
+    withr::defer(daemons(0), envir = envir)
+  }
+}
+
 getConfLevel <- function(level, fan) {
   if (fan) {
     seq(51, 99, by = 3)
@@ -14,7 +21,12 @@ get_seed <- function() {
   seed <- get0(".Random.seed", globalenv(), mode = "integer", inherits = FALSE)
   if (is.null(seed)) {
     runif(1L)
-    seed <- get0(".Random.seed", globalenv(), mode = "integer", inherits = FALSE)
+    seed <- get0(
+      ".Random.seed",
+      globalenv(),
+      mode = "integer",
+      inherits = FALSE
+    )
   }
   seed
 }