+++ tests/testthat/test-tbats.R
@@ -49,8 +49,14 @@ test_that("Test tbats() and forecasts", {
   ))
 })
 
-#test_that("Test tbats() with parallel", {
-# Tests will not run on Travis in parallel
-# expect_output(print(tbats(woolyrnq, num.cores = 1)), regexp = "TBATS")
-# expect_output(print(tbats(elecsales, num.cores = 1, use.trend = FALSE)), regexp = "BATS")
-#})
+test_that("tbats() parallel matches serial", {
+  y <- subset(wineind, end = 50)
+  serial <- tbats(y, use.parallel = FALSE)
+  parallel <- tbats(y, use.parallel = TRUE)
+  expect_equal(parallel$AIC, serial$AIC)
+  expect_equal(parallel$method, serial$method)
+})
+
+test_that("tbats() num.cores is deprecated", {
+  expect_snapshot(invisible(tbats(rep(1, 100), num.cores = 1, use.parallel = FALSE)))
+})